fix action bar test
